backend_old-python/requirements.txt
Django>=4.2
djangorestframework
channels
channels_redis
django-cors-headers
daphne
tsplib95
